requirements.txt
groq
fastapi
uvicorn
streamlit
python-dotenv
requests
pydantic